move huge similar _evalue logic into separate file
